refactor: [#12] view 생성 코드 정리
- spacing 등 중복 유닛 상수화
- 디렉토리 구분
- 줄간격 조정
