+++ iOS/issue-tracker/issue-tracker/Common/Colors.swift
@@ -9,7 +9,7 @@ import UIKit
 
 enum Colors {
     static let border = UIColor(named: "Border") ?? UIColor.systemGray4
-    static let backGround = UIColor(named: "BackGround") ?? UIColor.systemGray5
+    static let background = UIColor(named: "BackGround") ?? UIColor.systemGray5
     static let closeCarrot = UIColor(named: "CloseCarrot") ?? UIColor.orange
     static let deleteSunnyside = UIColor(named: "DeleteSunnyside") ?? UIColor.yellow
     static let openBrocoli = UIColor(named: "OpenBrocoli") ?? UIColor.purple
