--- iOS/issue-tracker/issue-tracker/UIImageExtension.swift
@@ -1,22 +1,0 @@
-//
-//  UIImageExtension.swift
-//  issue-tracker
-//
-//  Created by jinseo park on 6/8/21.
-//
-
-import UIKit
-
-extension UIImage
-{
-  func resizedImage(Size sizeImage: CGSize) -> UIImage?
-  {
-      let frame = CGRect(origin: CGPoint.zero, size: CGSize(width: sizeImage.width, height: sizeImage.height))
-      UIGraphicsBeginImageContextWithOptions(frame.size, false, 0)
-      self.draw(in: frame)
-      let resizedImage: UIImage? = UIGraphicsGetImageFromCurrentImageContext()
-      UIGraphicsEndImageContext()
-      self.withRenderingMode(.alwaysOriginal)
-      return resizedImage
-  }
-}
